# contracts/openapi/aihub-v1.yaml
...
openapi: 3.1.0
...
info:
  title: AIHub Platform API
  version: v1
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    内部 AI 资产管理平台 REST API 契约。统一成功响应为 ApiResponse，统一失败响应为 ApiError。
    Agent 不应依赖本地化 message 做分支判断，只依赖稳定的 code 与 retryable。
servers:
  - url: /api/v1
    description: 统一 API 前缀
...
security:
  - bearerAuth: []
paths:
...
  /system/dictionaries:
    get:
      operationId: listDictionaries
      summary: 查询字典类型
      tags: [taxonomy]
      x-implementation-status: planned-p0-b
      x-required-permissions: [dictionary:read]
      responses:
        '200':
          description: 字典类型列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/DictionaryTypeView' } }
        '403': { $ref: '#/components/responses/ErrorResponse' }
  /system/dictionaries/{dictCode}/items:
    parameters:
      - { name: dictCode, in: path, required: true, schema: { type: string } }
    get:
      operationId: listDictionaryItems
      summary: 查询字典项
      tags: [taxonomy]
      x-implementation-status: planned-p0-b
      x-required-permissions: [dictionary:read]
      responses:
        '200':
          description: 字典项列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/DictionaryItemView' } }
        '403': { $ref: '#/components/responses/ErrorResponse' }
    post:
      operationId: createDictionaryItem
      summary: 新增字典项
      tags: [taxonomy]
      x-implementation-status: planned-p0-b
      x-required-permissions: [dictionary:manage]
      x-audit-event: DICTIONARY_ITEM_CREATED
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDictionaryItemRequest' }
      responses:
        '201':
          description: 字典项创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DictionaryItemView' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /system/dictionaries/{dictCode}/items/{itemCode}:
    patch:
      operationId: updateDictionaryItem
      summary: 更新、启用或停用字典项
      tags: [taxonomy]
      x-implementation-status: planned-p0-b
      x-required-permissions: [dictionary:manage]
      x-audit-event: DICTIONARY_ITEM_UPDATED
      parameters:
        - { name: dictCode, in: path, required: true, schema: { type: string } }
        - { name: itemCode, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDictionaryItemRequest' }
      responses:
        '200':
          description: 字典项更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DictionaryItemView' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        平台签发的短期访问 JWT。JWT 使用非对称签名和 kid；资源级权限由后端实时判断，
        不以 JWT 内角色替代 RBAC/ACL/资源状态校验。
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 当前 Principal 范围内唯一的写请求幂等键。
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    ErrorResponse:
      description: 统一失败响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiResponse:
      type: object
      description: 统一成功响应包装
      required:
        - requestId
        - traceId
        - timestamp
      properties:
        data:
          description: 业务数据载荷，结构由具体接口定义
          type: [object, array, string, number, integer, boolean, 'null']
        requestId:
          type: string
          description: 请求标识，回显 X-Request-Id
          example: req_01J0000000000000000000000
        traceId:
          type: string
          description: 分布式追踪标识
          example: 4bf92f3577b34da6a3ce929d0e0e4736
        timestamp:
          type: string
          format: date-time
          description: 服务端 UTC 时间戳
          example: '2026-06-29T08:00:00Z'
    ApiError:
      type: object
      description: 统一失败响应
      required:
        - code
        - message
        - retryable
        - requestId
        - traceId
      properties:
        code:
          type: string
          description: 稳定的语义错误码，分支判断只依赖此字段
          example: ASSET_VERSION_CONFLICT
        message:
          type: string
          description: 默认（可本地化）错误消息，不用于分支判断
          example: Version already exists
        i18nKey:
          type: string
          description: 国际化文案 Key
          example: error.asset.versionConflict
        details:
          type: object
          description: 附加上下文，结构由错误码定义
          additionalProperties: true
        retryable:
          type: boolean
          description: 是否可重试；仅当为 true 时 Agent 才应按退避策略重试
          example: false
        requestId:
          type: string
          example: req_01J0000000000000000000000
        traceId:
          type: string
          example: 4bf92f3577b34da6a3ce929d0e0e4736
...
    TaxonomyStatus:
      type: string
      enum: [ACTIVE, DISABLED]
    DictionaryTypeView:
      type: object
      required: [dictCode, i18nKey, version, status]
      properties:
        dictCode: { type: string }
        i18nKey: { type: string }
        version: { type: integer, minimum: 1 }
        status: { $ref: '#/components/schemas/TaxonomyStatus' }
    DictionaryItemView:
      type: object
      required: [dictCode, itemCode, i18nKey, sortOrder, status, version]
      properties:
        dictCode: { type: string }
        itemCode: { type: string }
        i18nKey: { type: string }
        sortOrder: { type: integer }
        status: { $ref: '#/components/schemas/TaxonomyStatus' }
        version: { type: integer, minimum: 1 }
    CreateDictionaryItemRequest:
      type: object
      additionalProperties: false
      required: [itemCode, i18nKey]
      properties:
        itemCode:
          type: string
          pattern: '^[a-zA-Z0-9][a-zA-Z0-9._-]{0,63}$'
        i18nKey: { type: string }
        sortOrder: { type: integer, default: 0 }
    UpdateDictionaryItemRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        i18nKey: { type: string }
        sortOrder: { type: integer }
        status: { $ref: '#/components/schemas/TaxonomyStatus' }
        expectedVersion: { type: integer, minimum: 1 }
...